outsource logic into proxyGrafanaDatasource
