+++ tests/fast-integration/upgrade-test/upgrade-test-tests.js
@@ -2,7 +2,7 @@ const {checkServiceInstanceExistence} = require('./fixtures/helm-broker');
 
 const {printRestartReport, getContainerRestartsForAllNamespaces} = require('../utils');
 const {loggingTests} = require('../logging');
-const {monitoringTests} = require('../monitoring');
+const {monitoringTests, resetGrafanaProxy} = require('../monitoring');
 const {tracingTests} = require('../tracing');
 const {checkInClusterEventDelivery,
   checkFunctionResponse,
@@ -44,4 +44,5 @@ describe('Upgrade test tests', function() {
   monitoringTests();
   loggingTests();
   tracingTests(mockNamespace, testNamespace);
+  resetGrafanaProxy();
 });
